test: make Azure CAS local e2e invocation work on 1.35 (#41)
* feat: e2e test framework improvements for local + CI execution

Reorganizes the Azure CAS e2e test infrastructure to support both
local developer execution and CI (Prow), following patterns from
karpenter-provider-azure.

Changes:
- New shared Environment struct (pkg/environment/) with K8s + VMSS
  clients and optional Helm deploy
- Moved scale-up test into suites/scaleup/, uses shared environment
- Deleted old test/e2e/ (replaced by suites/ + pkg/environment/)
- Updated Makefile with local targets (e2etests, setup-cluster,
  deploy-local) and preserved CI target
- Added devcontainer.json for Azure CAS dev (Go 1.25)
- Added agent skill doc for e2e test workflows
- aks-dev-deploy.sh creates settings.json configmap for fork
- cluster-autoscaler-vmss-wi-dynamic.yaml.tpl adds e2e-friendly
  CAS args and configmap volume mount

Review feedback addressed:
- Require AZURE_RESOURCE_GROUP in aks-dev-deploy.sh, fail fast
- Add breadcrumb @echo to Makefile targets (setup-cluster,
  deploy-local, e2etests)
- Fix TEST_SUITE default so ./suites/... works correctly
- AllVMSSStable now checks every node is Ready (not just one)
- NewEnvironment validates required params early with clear errors

* azure: reorganize e2e test framework for per-suite execution

Restructure Azure CAS e2e tests from a single e2e/ directory into a
suites/ + pkg/environment/ layout. This enables per-suite execution,
shared test infrastructure, and supports both local developer and CI
(Prow) workflows.

- Extract shared Environment struct with K8s/VMSS clients into
  pkg/environment/environment.go
- Move scale-up test to suites/scaleup/suite_test.go
- Make Helm deployment optional (CI deploys via Helm flags, local
  assumes CAS is already running)
- Add local developer targets to Makefile (e2etests, setup-cluster,
  deploy-local, list-suites, validate-env)
- Update CI target to use suites/ layout

(cherry picked from commit 592f370ec120dc43da8fb05832d0e2cf4f5a51d5)

* feat: add devcontainer for development environment

Adds a devcontainer configuration with:
- Go 1.25 base image (mcr.microsoft.com/devcontainers/go:1.25)
- Docker-in-Docker for container image builds
- Azure CLI, kubectl, Helm, minikube for AKS workflows
- GitHub CLI for PR workflows
- libseccomp-dev system dependency
- Pre-warmed Go module caches for cluster-autoscaler and VPA
- CI verification tools (golint, godep, misspell) and kind pre-installed
- Go extension with CI-matching test and lint settings

(cherry picked from commit 0d974cf2813e53f5dae96bbdebe89f96ddd3394d)

---------
